+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept2D/NearestAreaScanWith2D.kt
@@ -2,7 +2,7 @@ package geometry.concept2D
 
 import elements.Coordinates
 import logic.GameField
-import logic.OneMoveProcessing
+import geometry.OneMoveProcessing
 import logic.PlayerProvider
 import utilities.Log
