+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept3D/NearestAreaScanWith3D.kt
@@ -2,7 +2,7 @@ package geometry.concept3D
 
 import elements.Coordinates
 import logic.GameField
-import logic.OneMoveProcessing
+import geometry.OneMoveProcessing
 import logic.PlayerProvider
 import utilities.Log
