+++ Attt_LocalEngine_Kotlin/src/main/kotlin/logic/GameField.kt
@@ -6,6 +6,7 @@ import constants.SYMBOL_FOR_ABSENT_MARK
 import constants.SYMBOL_FOR_DIVIDER
 import constants.SYMBOL_FOR_NEW_LINE
 import elements.*
+import geometry.OneMoveProcessing
 import geometry.conceptXY.NearestAreaScanWithXY
 import publicApi.AtttPlayer
 import utilities.Log
